Project Bones Created
Created a main file with a main class to run the application code.  Also created a test txt file for each problem type.  only uncomment the test file for what you are testing to test.  Also created a file for each problem type and import them all to the main.  Each py file is named after which problem it will solve and the idea with the current code is to do the calculations in each seperate py file and return a formatted string to print to the file.  Also created the readwrite py file that parses a formatted txt file into an array and will also create a time stamped txt file that holds the returned value of which ever py executes.